% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cascade.R
\name{create_cascade}
\alias{create_cascade}
\title{Create plot of selected cascades}
\usage{
create_cascade(msd_df)
}
\arguments{
\item{msd_df}{MSD data frame from either a Genie or Panorama MSD}
}
\description{
`create_cascade` is a wrapper function that lets a user select between
13 different cascade plots. Valid entries must fall between 1 and 10. The
function will plot the selected cascade and return a list of data underlying
the plot.
}
\examples{
\dontrun{
# Create a standard cascade plot
create_cascade(df)
1
}
}
