% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keep_ind}
\alias{keep_ind}
\title{Name of annotation indicators to include on plots}
\format{
A list of all snapshot indicators
\describe{
  \item{keep_ind}{indicator list}
}
}
\usage{
data(keep_ind)
}
\description{
List of indicators .
}
\keyword{datasets}
