% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cascade.R
\name{plot_cascade}
\alias{plot_cascade}
\title{Plot Cascade
`plot_cascade` is a ggplot2-based plot that arranges, annotates and
plots data from the 90-90-90 cascade. Four inputs are required to create
a valid plot.}
\usage{
plot_cascade(.data, df_annot, cscd_num, p_title)
}
\arguments{
\item{.data}{dataframe that is the result of `return_cascade` and `assign_cscd_colors`}

\item{df_annot}{annotation dataframe created from the `create_annotations` function}

\item{cscd_num}{cascade number selected by the user}

\item{p_title}{plot title derived from the `cscd_num`}
}
\description{
Plot Cascade
`plot_cascade` is a ggplot2-based plot that arranges, annotates and
plots data from the 90-90-90 cascade. Four inputs are required to create
a valid plot.
}
