% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_cascade.R
\name{return_cascade}
\alias{return_cascade}
\title{Title}
\usage{
return_cascade(df, cscd_num)
}
\arguments{
\item{df}{MER structured data set for a given population and geography}

\item{cscd_num}{cascade selection from user}
}
\description{
Title
}
