% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annotations.R
\name{create_annotations}
\alias{create_annotations}
\title{Title}
\usage{
create_annotations(.data, cscd_df)
}
\arguments{
\item{.data}{munged data frame resulting from `create_cascade`}

\item{cscd_df}{cascade value from the `create_cascade`}
}
\description{
Title
}
